chore: revert prisma splitting
