+++ src/server/components/auth/__tests__/authenticated-user.test.ts
@@ -1,7 +1,7 @@
 import AuthenticatedUser from "../authenticated-user";
 import { prisma } from "../../../models/db/__mocks__/prisma-client";
 
-jest.mock("../../../../shared/prisma");
+jest.mock("./../../../models/db/prisma-client");
 
 
 let user;
