+++ src/server/models/listItem/__tests__/listItem.update.test.ts
@@ -2,7 +2,7 @@ import * as location from "../../../../server/services/location";
 import * as listItem from "./../listItem";
 import { prisma } from "../../../models/db/__mocks__/prisma-client";
 
-jest.mock("../../../../shared/prisma");
+jest.mock("./../../../models/db/prisma-client");
 
 test("throws when the requested id does not exist", async () => {
   // @ts-expect-error
